fix: fixing imports in tests and repackaging tests
