+++ app/src/androidTest/java/com/github/studydistractor/sdp/procrastinationActivity/ProcrastinationActivityActivityTest.kt
@@ -1,12 +1,15 @@
 package com.github.studydistractor.sdp.procrastinationActivity
 
+import android.app.Activity
+import android.content.Context
 import android.content.Intent
+import androidx.compose.ui.test.junit4.ComposeTestRule
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
-import com.github.studydistractor.sdp.launch
-import com.github.studydistractor.sdp.register.RegisterActivity
+import androidx.test.core.app.ActivityScenario
+import androidx.test.core.app.ApplicationProvider
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Assert.fail
@@ -15,6 +18,25 @@ import org.junit.Rule
 import org.junit.Test
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
+
+/**
+ * Uses a [ComposeTestRule] created via [createEmptyComposeRule] that allows setup before the activity
+ * is launched via [onBefore]. Assertions on the view can be made in [onAfterLaunched].
+ *
+ * source : https://stackoverflow.com/questions/68267861/add-intent-extras-in-compose-ui-test
+ */
+inline fun <reified A: Activity> ComposeTestRule.launch(
+    onBefore: () -> Unit = {},
+    intentFactory: (Context) -> Intent = { Intent(ApplicationProvider.getApplicationContext(), A::class.java) },
+    onAfterLaunched: ComposeTestRule.() -> Unit
+) {
+    onBefore()
+
+    val context = ApplicationProvider.getApplicationContext<Context>()
+    ActivityScenario.launch<A>(intentFactory(context))
+
+    onAfterLaunched()
+}
 
 @HiltAndroidTest
 class ProcrastinationActivityActivityTest {
